refactor(structure): reorganize project folder structure
