+++ src/components/sections/HeroSlider/HeroSlider.tsx
