+++ src/components/sections/LareiraSection/LareiraSection.tsx
