+++ src/data/HeroSliderData.ts
